updt

+++ WeChip.DomainModel/Models/DeliveryClientAddressModel.cs
@@ -9,7 +9,7 @@ namespace WeChip.DomainModel.Models
     {
         public string CEP { get; set; }
         public string Street { get; set; }
-        public byte NumberAddress { get; set; }
+        public int NumberAddress { get; set; }
         public string Complement { get; set; }
         public string Neighbourhood { get; set; }
         public string City { get; set; }
